requirement.txt
instructor==1.7.2
google-generativeai==0.8.4
llama-cloud-services==0.6.1
streamlit==1.42.2